Check cert path.

+++ pkg/mac/build-cpp-deps-lib.sh
@@ -64,10 +64,16 @@ if [ ! -f openssl-OpenSSL_${OPENSSL_VERSION_UNDERSCORE}.done ]; then
         else
           PLATFORM=darwin64-x86_64-cc
         fi
-        ./Configure --prefix=$PREFIX $SSL_DIR --openssldir=/etc/ssl no-shared no-unit-test $PLATFORM
+        ./Configure --prefix=$PREFIX --openssldir=/etc/ssl no-shared no-unit-test $PLATFORM
         make -j8
         make install_sw
     popd
+
+    OPENSSL_DIR=$(./install/bin/openssl version -d | sed 's/\"//g')
+    if [[ $OPENSSL_DIR =~ "\/etc\/ssl" ]]; then
+        echo "The openssl dir is not set as expected: " $OPENSSL_DIR
+        exit 1
+    fi
 
     rm -rf OpenSSL_${OPENSSL_VERSION_UNDERSCORE}.tar.gz openssl-OpenSSL_${OPENSSL_VERSION_UNDERSCORE}
     touch openssl-OpenSSL_${OPENSSL_VERSION_UNDERSCORE}.done
@@ -179,6 +185,12 @@ if [ ! -f curl-${CURL_VERSION}.done ]; then
               --host=$ARCH-apple-darwin
       make -j16 install
     popd
+
+    CERT_PATH=$(./install/bin/curl -v https://example.com 2>&1 | grep CAfile)
+    if [[ $CERT_PATH =~ "\/etc\/ssl" ]]; then
+        echo "The certification path is not set as expected: " $CERT_PATH
+        exit 1
+    fi
 
     rm -rf curl-${CURL_VERSION} curl-${CURL_VERSION}.tar.gz
     touch curl-${CURL_VERSION}.done
